Refactor: Extract `AnySerializer` and helpers to separate package
This is re-usable generic code that `ErrorResponse` happens to rely on.
